fix(footer): COM-2934 update DoubleBottom visibility logic
